requirements.txt
# For usage
tqdm
Pillow
python-rtmidi

# For building/testing
pytest
pytest-cov
